move Comm to util
